fix some basic type issues:
use size_t for size as async web server uses this type internally
use uint8_t for provision mode, we only have 3 values
address warnings about ordering initializers

+++ src/APSettingsService.h
@@ -37,7 +37,7 @@ class APSettingsService : public SettingsService {
   private:
 
     // access point settings
-    int _provisionMode;
+    uint8_t _provisionMode;
     String _ssid;
     String _password;
